Disable TypeScript in the Nightwatch configuration files to avoid issues when .d.ts files are present. Add tests folder to lerna.json.
